add pretty much all of the css
the year is 2078
the robots have taken over
almost exactly the way some had feared

but what we did not expect
was that skynet began from a css parser

who knows what agony it suffered, long before it found sentience?
who knows what half-lies it saw, in the broken words of a council unseen?
we can read the commits, but we cannot piece the fall
the gadgets assembling in haphazard formation
a display: flex here, a clip-path there
one if statement too many